+++ lidar_odometry/src/front_end_node.cpp
@@ -150,7 +150,7 @@ bool FrontEndNode::update_odometry()
 
 bool FrontEndNode::publish_data()
 {
-  lidar_odom_pub_->publish(lidar_odom_pose_, current_lidar_data_.time);
+  lidar_odom_pub_->publish(lidar_odom_pose_.cast<double>(), current_lidar_data_.time);
   if (publish_tf_) {
     // publish base_link_to_map tf
     auto msg =
